Booth Asset
Created a script for the Booth Asset. The player starts the scene sitting at the booth. If they point the right controller down at the ground and press "A", they are free to move. If the point the right controller at the booth and press "A", they sit at the booth.