Added functionality to AI script for food to be delivered
In the event system make sure to remove standalone input module and replace with XR UI Input Module or else the menu buttons will not work with XR Rays.

Added to the AI script, now when a food item is ordered the AI will go to the table and "drop off" the food for the player, when they decide to pick it up it will take them to a new scene.

Menu's buttons are all functions from the Server AI script, they are for each specific food item so each button needs to be set to the appropriate food.

The food items have the XR Grab Interactable and under interactable Events they have the function to change scenes in select. So when the player picks up the item it will change scene.