Added Teleportation zone and functionality, added on to XR origin for teleportation
Added Teleportation rays that only show when teleportation is set to on and the player holds the trigger. Then it teleports when they release them. Added invisible plane to diner for teleportation as using the dinner floor also let players teleport onto the roof. Added teleportation zones as well as made it so they dont show up on the UI with the teleportation Rays.