Changed Teleportation, Added Menu attach point and changed Events to Dont destroy on load
Changed Teleportation to be using the left thumb stick, line appears when held down and you teleport when let go. There is also a reticle I made real fast, it is a bright color to see it easily, we can remove or change how it looks anytime.
Added an attach point to the menu so it looks better when holding.
Lastly, I added a "Dont Destroy" script to the event script, as it needs to be the same event as the one it loaded in with in the menu or else you will not be able to interact with UI elements in the Hub world scene.