Fixed Bug With Booth.cs
When exiting the booth it would sometimes not let you move and you would be floating, i was able to fix it. I also changed one line on the PlaywerSpawnPoint.cs to make the players rotation the same as the spawn point so they are not looking at the wall when they spawn.